requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
requests==2.26.0
icalendar==6.1.1
python-dateutil==2.8.2
gunicorn==21.2.0